requirements.txt
numpy
scikit-image
scipy
sklearn
smplx==0.1.28
tqdm
yacs
numba
opencv-python
tensorboardx
filterpy
cython
chumpy
Pillow
trimesh
pyrender
matplotlib
json_tricks
torchgeometry
einops
joblib
boto3
requests
easydict
pycocotools
plyfile
mmcv-full==1.5.0